Evaluate debugging scripts on bigger pool of data
